add error handling for the getWeather fnc.

+++ data/migrations/20210402203703_savedweather.js
@@ -3,7 +3,7 @@ exports.up = function (knex) {
     table.increments().primary();
     table.string("city").notNullable();
     table.decimal("temp").notNullable();
-    table.decimal("humidity").notNullable();
+    table.integer("humidity").notNullable();
     table.timestamps(true, true);
   });
 };
